Rename game->games and user->users
