+++ web-app/src/games/games.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { GameController } from './game.controller';
-import { GameService } from './game.service';
+import { GameController } from './games.controller';
+import { GameService } from './games.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Game } from './game.entity';
 import { getConnectionOptions } from 'typeorm';
-import { User } from '../user/user.entity';
+import { User } from '../users/user.entity';
 
 @Module({
   imports: [
