+++ web-app/src/localplay/games/localplay.games.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { GamesController } from './games.controller';
+import { GamesController } from './localplay.games.controller';
 
 @Module({
   controllers: [GamesController],
